refactor/add folder for api's and move WS helpers to api's folder
