+++ client/src/components/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { sendMessageToServer, getBotMessage } from '../api';
+import { sendMessageToServer, getBotMessage } from '../API/WShelpers';
 import '../App.css';
 
 class App extends Component {
